move into browser router provider
