+++ src/components/Layout/Main.tsx
@@ -2,9 +2,9 @@ import { Box, Container, ContainerProps, HStack, Stack } from '@chakra-ui/layout
 import { useColorModeValue } from '@chakra-ui/react'
 import { useViewportScroll } from 'framer-motion'
 import { ReactNode, useEffect, useRef, useState } from 'react'
-import { useAppRoutes } from 'Routes/Routes'
 import { Breadcrumbs } from 'components/Breadcrumbs/Breadcrumbs'
 import { NestedMenu } from 'components/NestedMenu/NestedMenu'
+import { useBrowserRouter } from 'context/BrowserRouterProvider/BrowserRouterProvider'
 
 import { Page } from './Page'
 
@@ -14,7 +14,7 @@ export type MainProps = {
 
 export const Main: React.FC<MainProps> = ({ children, titleComponent, ...rest }) => {
   const ref = useRef<HTMLDivElement>(null)
-  const { currentRoute } = useAppRoutes()
+  const { currentRoute } = useBrowserRouter()
   const bg = useColorModeValue('white', 'gray.800')
   const borderColor = useColorModeValue('gray.100', 'gray.750')
   const [y, setY] = useState(0)
