+++ src/context/BrowserRouterProvider/BrowserRouterProvider.tsx
@@ -1,6 +1,11 @@
 import { History, Location } from 'history'
+import { union } from 'lodash'
 import React, { useContext, useMemo } from 'react'
-import { useHistory, useLocation, useParams } from 'react-router-dom'
+import { matchPath, useHistory, useLocation, useParams } from 'react-router-dom'
+import { generateAppRoutes } from 'Routes/helpers'
+import { Route as NestedRoute } from 'Routes/helpers'
+import { routes } from 'Routes/Routes'
+import { usePlugins } from 'context/PluginProvider/PluginProvider'
 import { useQuery } from 'hooks/useQuery/useQuery'
 
 type BrowserRouterContextProps<Q, P> = {
@@ -8,6 +13,8 @@ type BrowserRouterContextProps<Q, P> = {
   history: History
   params: P
   query: Q
+  appRoutes: NestedRoute[]
+  currentRoute: NestedRoute | void
 }
 
 const BrowserRouterContext = React.createContext<BrowserRouterContextProps<any, any> | null>(null)
@@ -27,15 +34,26 @@ export function BrowserRouterProvider({ children }: BrowserRouterProviderProps) 
   const history = useHistory()
   const params = useParams()
   const query = useQuery()
+  const { routes: pluginRoutes } = usePlugins()
+
+  const appRoutes = useMemo(() => {
+    return generateAppRoutes(union(pluginRoutes, routes))
+  }, [pluginRoutes])
+
+  const currentRoute = useMemo(() => {
+    return appRoutes.find(e => matchPath(location.pathname, { path: e.path, exact: true }))
+  }, [appRoutes, location.pathname])
 
   const router = useMemo(
     () => ({
       history,
       location,
       params,
-      query
+      query,
+      appRoutes,
+      currentRoute
     }),
-    [query, params, location, history]
+    [history, location, params, query, appRoutes, currentRoute]
   )
 
   return <BrowserRouterContext.Provider value={router}>{children}</BrowserRouterContext.Provider>
